+++ src/components/pages/Main/CategoriesSection/CategoriesSection.css
