+++ src/components/pages/RecruitingDetail/detail/RecruitingHeader.jsx
