+++ src/components/pages/RecruitingDetail/detail/RecruitingInfoSection.css
