+++ src/components/pages/RecruitingDetail/detail/RecruitingInfoSection.jsx
