+++ src/components/pages/RecruitingDetail/detail/index.js
