+++ src/components/pages/RecruitingDetail/index.js
