+++ src/components/pages/RecruitingDetail/stats/StatsSection.css
