+++ src/components/pages/RecruitingDetail/stats/index.js
