--- src/components/sections/index.js
@@ -1,4 +1,0 @@
-// Section Components
-export { default as CategoriesSection } from './CategoriesSection/CategoriesSection';
-export { default as ContactSection } from './ContactSection/ContactSection';
-export { default as InfoSection } from './InfoSection/InfoSection';
